sidebar, home page modified, genres crud, age group crud, blogs operations

+++ src/app/views/age-group/age-group.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, TemplateRef } from "@angular/core";
 import { GenreService } from "src/app/services/genre/genre.service";
 import { Router } from "@angular/router";
-import { BsModalRef, BsModalService } from "ngx-bootstrap/modal";
+
 import { AgeGroupService } from "src/app/services/age-group/age-group.service";
 
 @Component({
@@ -12,24 +12,16 @@ import { AgeGroupService } from "src/app/services/age-group/age-group.service";
 export class AgeGroupComponent implements OnInit {
   data: Array<any> = [];
   genreAll: boolean = true;
-  isAddGenre: boolean = false;
+  isAddAge: boolean = false;
   isAddImage: boolean = false;
 
   from = "";
   to = "";
   topicMedia = [];
-  modalRef: BsModalRef;
-  Modelconfig = {
-    backdrop: true,
-    ignoreBackdropClick: true,
-    class: "modelWidth",
-  };
+  id;
+  editable: boolean = false;
 
-  constructor(
-    private service: AgeGroupService,
-    private router: Router,
-    private modalService: BsModalService
-  ) {}
+  constructor(private service: AgeGroupService, private router: Router) {}
 
   ngOnInit() {
     const route = this.router.url.split("/").pop();
@@ -46,11 +38,11 @@ export class AgeGroupComponent implements OnInit {
     this.router.navigateByUrl("/publications/" + url);
   }
 
-  enableAddGenre() {
-    this.isAddGenre = !this.isAddGenre;
+  enableAddAge() {
+    this.isAddAge = !this.isAddAge;
   }
 
-  addGenre() {
+  addAge() {
     if (this.from == "" || this.to == "") {
       alert("All the fields are required");
       return;
@@ -61,8 +53,23 @@ export class AgeGroupComponent implements OnInit {
         ageRange: age,
       })
       .subscribe((_response) => {
-        console.log(_response);
         alert(_response.body.message);
+        this.from = "";
+        this.to = "";
+        this.getAgeGroups();
+      });
+  }
+
+  updateAge() {
+    const age = `${this.from},${this.to}`;
+    this.service
+      .put(this.id, {
+        ageRange: age,
+      })
+      .subscribe((_response) => {
+        alert(_response.body.message);
+        this.from = "";
+        this.to = "";
         this.getAgeGroups();
       });
   }
@@ -80,24 +87,13 @@ export class AgeGroupComponent implements OnInit {
     });
   }
 
-  openModal(template: TemplateRef<any>, imgs) {
-    this.topicMedia = imgs;
-    this.modalRef = this.modalService.show(template, this.Modelconfig);
-  }
-
-  decline() {
-    this.modalRef.hide();
-    this.topicMedia = [];
-    this.isAddGenre = false;
-    this.isAddImage = false;
-  }
-
-  deleteGenre(id) {
-    this.service.delete(id).subscribe((_response) => {
-      alert(_response.body.message);
-      this.getAgeGroups();
-      // this.topicMedia = _response.body.data;
-    });
+  deleteAge(id) {
+    if (confirm("Are you sure want to delete this age group")) {
+      this.service.delete(id).subscribe((_response) => {
+        alert(_response.body.message);
+        this.getAgeGroups();
+      });
+    }
   }
 
   deleteOneImg(id, img) {
@@ -113,5 +109,13 @@ export class AgeGroupComponent implements OnInit {
   }
   enableAddImage() {
     this.isAddImage = !this.isAddImage;
+  }
+
+  editAge(row) {
+    this.from = row.ageRange[0];
+    this.to = row.ageRange[1];
+    this.id = row._id;
+    this.editable = true; // true for edit operation
+    this.isAddAge = true;
   }
 }
